+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/channel/member/internal/MemberEntity.kt
@@ -50,4 +50,10 @@ internal data class MemberEntity(
 
     /** The user's channel-level role. */
     val channelRole: String? = null,
+
+    /** If the notifications are muted for this user in the channel. */
+    val notificationsMuted: Boolean? = null,
+
+    /** The user's status. */
+    val status: String? = null,
 )